社交中心+后端框架

+++ music-client/src/views/setting/Setting.vue
@@ -2,11 +2,14 @@
   <div class="setting">
     <h1>设置</h1>
     <el-tabs tab-position="left">
-      <el-tab-pane label="个人资料" class="content">
+      <el-tab-pane v-if="!isAdmin" label="个人资料" class="content">
         <Personal-data></Personal-data>
       </el-tab-pane>
-      <el-tab-pane label="更改密码" class="content">
+      <el-tab-pane v-if="!isAdmin" label="更改密码" class="content">
         <Password></Password>
+      </el-tab-pane>
+      <el-tab-pane v-if="isAdmin" label="管理员改密" class="content">
+        <AdminPassword></AdminPassword>
       </el-tab-pane>
       <el-tab-pane label="账号和安全" class="content">
         <el-button type="danger" :icon="Delete" @click="cancelAccount">注销账号</el-button>
@@ -20,6 +23,7 @@ import { defineComponent, getCurrentInstance, computed, reactive } from "vue";
 import { Delete } from "@element-plus/icons-vue";
 import PersonalData from "./PersonalData.vue";
 import Password from "./Password.vue";
+import AdminPassword from "./AdminPassword.vue";
 import { HttpManager } from "@/api";
 import { useStore } from "vuex";
 import mixin from "@/mixins/mixin";
@@ -29,6 +33,7 @@ export default defineComponent({
   components: {
     PersonalData,
     Password,
+    AdminPassword,
   },
   setup() {
     const { proxy } = getCurrentInstance();
@@ -36,20 +41,31 @@ export default defineComponent({
     const { routerManager } = mixin();
 
     const userId = computed(() => store.getters.userId);
-
+    const isAdmin = computed(() => {
+      return Boolean(store.getters.isAdmin) || localStorage.getItem("cm_isAdmin") === "true";
+    });
     async function cancelAccount() {
-      const result = (await HttpManager.deleteUser(userId.value)) as ResponseBody;
-      (proxy as any).$message({
-        message: result.message,
-        type: result.type,
-      });
+      if (!isAdmin.value) {
+        const result = (await HttpManager.deleteUser(userId.value)) as ResponseBody;
+        (proxy as any).$message({
+          message: result.message,
+          type: result.type,
+        });
+      } else {
+        const result = (await HttpManager.adminLogout()) as ResponseBody;
+        (proxy as any).$message({
+          message: result.message,
+          type: result.type,
+        });
+      }
       routerManager(RouterName.SignIn, { path: RouterName.SignIn });
       proxy.$store.commit("setToken", false);
+      proxy.$store.commit("clearUser");
     }
-
     return {
       Delete,
       cancelAccount,
+      isAdmin,
     };
   },
 });
